Finish the TexturedMeshes (indexed and non-indexed)
  - Feature was implemented
  - Prototype was removed (HandleIndexing)
  - Test were added
  - Documentation was updated
