+++ src/main/scala/org/sorted/chaos/wavefront/mesh/TexturedMesh.scala
@@ -8,23 +8,23 @@ import org.sorted.chaos.wavefront.reader.Wavefront
   * @param vertices the array of vertices for the VertexBufferObject
   * @param textures the array of UV coordinates for the VertexBufferObject
   */
-final case class SimpleTexturedMesh(vertices: Array[Float], textures: Array[Float])
+final case class TexturedMesh(vertices: Array[Float], textures: Array[Float])
 
-object SimpleTexturedMesh extends Mesh {
+object TexturedMesh extends Mesh {
 
-  private def empty = SimpleTexturedMesh(Array.emptyFloatArray, Array.emptyFloatArray)
+  private def empty = TexturedMesh(Array.emptyFloatArray, Array.emptyFloatArray)
 
-  def from(wavefront: Wavefront): SimpleTexturedMesh = {
+  def from(wavefront: Wavefront): TexturedMesh = {
     val wavefrontVertices  = wavefront.vertices
     val wavefrontTextures  = wavefront.textures
     val wavefrontTriangles = wavefront.triangles
 
-    wavefrontTriangles.foldLeft(SimpleTexturedMesh.empty) { (accumulator, triangle) =>
+    wavefrontTriangles.foldLeft(TexturedMesh.empty) { (accumulator, triangle) =>
       {
         val vertices = getVerticesOfTriangle(triangle, wavefrontVertices)
         val textures = getTexturesOfTriangle(triangle, wavefrontTextures)
 
-        SimpleTexturedMesh(
+        TexturedMesh(
           vertices = accumulator.vertices ++ vertices,
           textures = accumulator.textures ++ textures
         )
